requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.0
pyyaml==6.0.1
openai==1.12.0
python-dateutil==2.8.2
tqdm==4.66.1
retrying==1.3.4
fake-useragent==1.4.0
